lastpush

+++ ft_printf.h
@@ -6,16 +6,16 @@
 /*   By: palaca <palaca@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/07 20:11:07 by palaca            #+#    #+#             */
-/*   Updated: 2024/11/07 20:42:33 by palaca           ###   ########.fr       */
+/*   Updated: 2024/11/08 02:12:24 by palaca           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef FT_PRINTF_H
 # define FT_PRINTF_H
 
-#include <unistd.h>
-#include <stdlib.h>
-#include <stdarg.h>
+# include <unistd.h>
+# include <stdlib.h>
+# include <stdarg.h>
 
 int	ft_control(char c, va_list l);
 int	ft_printf(const char *s, ...);
@@ -23,7 +23,7 @@ int	ft_putchar(char a);
 int	ft_putstr(char *str);
 int	ft_putnbr(long nb);
 int	ft_putunbr(unsigned int nb);
-int	ft_puthex(unsigned int nb, char x);
+int	ft_puthex(unsigned long nb, char x);
 int	ft_putpoint(unsigned long nb);
 
 #endif
